splitting promise
